Move to release 2.6.0

+++ opentreemap/treemap/views/map_feature.py
@@ -7,15 +7,18 @@ import json
 import hashlib
 from functools import wraps
 
-from django.http import Http404
-from django.template import RequestContext
+from django.template import RequestContext, TemplateDoesNotExist
+from django.template.loader import get_template
 from django.shortcuts import get_object_or_404, render_to_response
 from django.core.exceptions import ValidationError
 from django.conf import settings
 from django.db import transaction
 from django.contrib.gis.geos import Point, MultiPolygon, Polygon
+from django.contrib.gis.db.models import GeometryField
 
 from opentreemap.util import dotted_split
+from treemap.lib.hide_at_zoom import (update_hide_at_zoom_after_move,
+                                      update_hide_at_zoom_after_delete)
 
 from treemap.units import Convertible
 from treemap.models import (Tree, Species, MapFeature,
@@ -26,21 +29,28 @@ from treemap.images import get_image_from_request
 from treemap.lib.photo import context_dict_for_photo
 from treemap.lib.object_caches import udf_defs
 from treemap.lib.map_feature import (get_map_feature_or_404,
+                                     raise_non_instance_404,
                                      context_dict_for_plot,
                                      context_dict_for_resource,
                                      context_dict_for_map_feature)
+from treemap.lib.perms import map_feature_is_deletable
+from treemap.views.misc import add_map_info_to_context
 
 
 def _request_to_update_map_feature(request, feature):
     request_dict = json.loads(request.body)
     feature, tree = update_map_feature(request_dict, request.user, feature)
 
+    ctx_fn = (context_dict_for_plot if feature.is_plot
+              else context_dict_for_resource)
+
     return {
         'ok': True,
         'geoRevHash': feature.instance.geo_rev_hash,
+        'universalRevHash': feature.instance.universal_rev_hash,
         'featureId': feature.id,
         'treeId': tree.id if tree else None,
-        'feature': context_dict_for_map_feature(request, feature),
+        'feature': ctx_fn(request, feature),
         'enabled': feature.instance.feature_enabled('add_plot'),
     }
 
@@ -68,64 +78,88 @@ def get_photo_context_and_errors(fn):
     return wrapper
 
 
-def map_feature_detail(request, instance, feature_id, render=False):
+def map_feature_detail(request, instance, feature_id,
+                       render=False, edit=False):
     feature = get_map_feature_or_404(feature_id, instance)
 
     ctx_fn = (context_dict_for_plot if feature.is_plot
               else context_dict_for_resource)
-    context = ctx_fn(request, feature)
+    context = ctx_fn(request, feature, edit=edit)
+    add_map_info_to_context(context, instance)
 
     if render:
         if feature.is_plot:
             template = 'treemap/plot_detail.html'
         else:
-            template = 'map_features/%s_detail.html' % feature.feature_type
-        return render_to_response(template, context, RequestContext(request))
+            app = feature.__module__.split('.')[0]
+            try:
+                template = '%s/%s_detail.html' % (app, feature.feature_type)
+                get_template(template)
+            except TemplateDoesNotExist:
+                template = 'treemap/resource_detail.html'
+        return render_to_response(template, context,
+                                  RequestContext(request))
     else:
         return context
 
 
-def render_map_feature_detail(*args, **kwargs):
-    return map_feature_detail(*args, render=True, **kwargs)
+def render_map_feature_detail(request, instance, feature_id, **kwargs):
+    return map_feature_detail(request, instance, feature_id, render=True,
+                              **kwargs)
+
+
+def context_map_feature_detail(request, instance, feature_id, **kwargs):
+    return map_feature_detail(request, instance, feature_id, render=False,
+                              **kwargs)
 
 
 def map_feature_photo_detail(request, instance, feature_id, photo_id):
-    photo = get_object_or_404(MapFeaturePhoto,
-                              pk=photo_id,
-                              map_feature=feature_id)
+    feature = get_map_feature_or_404(feature_id, instance)
+    photo = get_object_or_404(MapFeaturePhoto, pk=photo_id,
+                              map_feature=feature)
     return {'photo': context_dict_for_photo(request, photo)}
 
 
 def plot_detail(request, instance, feature_id, edit=False, tree_id=None):
     feature = get_map_feature_or_404(feature_id, instance, 'Plot')
-    return context_dict_for_plot(request, feature, edit, tree_id)
+    return context_dict_for_plot(request, feature, edit=edit, tree_id=tree_id)
 
 
 def render_map_feature_add(request, instance, type):
     if type in instance.map_feature_types[1:]:
-        template = 'map_features/%s_add.html' % type
-        return render_to_response(template, None, RequestContext(request))
+        app = MapFeature.get_subclass(type).__module__.split('.')[0]
+        try:
+            template = '%s/%s_add.html' % (app, type)
+            get_template(template)
+        except:
+            template = 'treemap/resource_add.html'
+        return render_to_response(template,
+                                  {'object_name': to_object_name(type)},
+                                  RequestContext(request))
     else:
-        raise Http404('Instance does not support feature type ' + type)
+        raise_non_instance_404(type)
 
 
 def add_map_feature(request, instance, type='Plot'):
-    feature = MapFeature.create(type, instance)
+    if type not in instance.map_feature_types:
+        raise_non_instance_404(type)
+    feature = MapFeature.get_subclass(type)(instance=instance)
     return _request_to_update_map_feature(request, feature)
 
 
-def update_map_feature_detail(request, instance, feature_id, type='Plot'):
-    feature = get_map_feature_or_404(feature_id, instance, type)
+def update_map_feature_detail(request, instance, feature_id):
+    feature = get_map_feature_or_404(feature_id, instance)
     return _request_to_update_map_feature(request, feature)
 
 
-def delete_map_feature(request, instance, feature_id, type='Plot'):
-    feature = get_map_feature_or_404(feature_id, instance, type)
-    try:
-        feature.delete_with_user(request.user)
-        return {'ok': True}
-    except ValidationError as ve:
-        return "; ".join(ve.messages)
+def delete_map_feature(request, instance, feature_id):
+    feature = get_map_feature_or_404(feature_id, instance)
+    if not map_feature_is_deletable(request.user.get_instance_user(instance),
+                                    feature):
+        raise ValidationError("map feature not deletable")
+    feature.delete_with_user(request.user)
+    update_hide_at_zoom_after_delete(feature)
+    return {'ok': True}
 
 
 @transaction.atomic
@@ -192,10 +226,10 @@ def update_map_feature(request_dict, user, feature):
         except ValidationError as e:
             return package_field_errors(thing._model_name, e)
 
-    old_location = feature.geom
-
     tree = None
 
+    rev_updates = ['universal_rev']
+    old_geom = feature.geom
     for (identifier, value) in request_dict.iteritems():
         split_template = 'Malformed request - invalid field %s'
         object_name, field = dotted_split(identifier, 2,
@@ -208,7 +242,7 @@ def update_map_feature(request_dict, user, feature):
                            if fdef.iscollection]
 
         if ((field in tree_udfc_names and
-             feature.current_tree() is None and
+             feature.safe_get_current_tree() is None and
              value == [])):
             continue
         elif object_name in feature_object_names:
@@ -216,7 +250,7 @@ def update_map_feature(request_dict, user, feature):
         elif object_name == 'tree' and feature.feature_type == 'Plot':
             # Get the tree or spawn a new one if needed
             tree = (tree or
-                    feature.current_tree() or
+                    feature.safe_get_current_tree() or
                     Tree(instance=feature.instance))
 
             # We always edit in display units
@@ -234,6 +268,13 @@ def update_map_feature(request_dict, user, feature):
 
         set_attr_on_model(model, field, value)
 
+        field_class = model._meta.get_field_by_name(field)[0]
+        if isinstance(field_class, GeometryField):
+            rev_updates.append('geo_rev')
+            rev_updates.append('eco_rev')
+        elif identifier in ['tree.species', 'tree.diameter']:
+            rev_updates.append('eco_rev')
+
     errors = {}
 
     if feature.fields_were_updated():
@@ -249,8 +290,10 @@ def update_map_feature(request_dict, user, feature):
             errors['mapFeature.geom'] = errors[feature.geom_field_name]
         raise ValidationError(errors)
 
-    if old_location is None or not feature.geom.equals_exact(old_location):
-        feature.instance.update_geo_rev()
+    if old_geom is not None and feature.geom != old_geom:
+        update_hide_at_zoom_after_move(feature, user, old_geom)
+
+    feature.instance.update_revs(*rev_updates)
 
     return feature, tree
 
@@ -264,14 +307,12 @@ def map_feature_hash(request, instance, feature_id, edit=False, tree_id=None):
     this function is wrapped around views that can take
     tree_id as an argument
     """
-
-    InstanceMapFeature = instance.scope_model(MapFeature)
-    base = get_object_or_404(InstanceMapFeature, pk=feature_id).hash
+    feature = get_map_feature_or_404(feature_id, instance)
 
     if request.user:
         pk = request.user.pk or ''
 
-    return hashlib.md5(base + ':' + str(pk)).hexdigest()
+    return hashlib.md5(feature.hash + ':' + str(pk)).hexdigest()
 
 
 @get_photo_context_and_errors
@@ -286,7 +327,10 @@ def rotate_map_feature_photo(request, instance, feature_id, photo_id):
         raise ValidationError('"degrees" must be a multiple of 90°')
 
     degrees = int(orientation)
-    mf_photo = get_object_or_404(MapFeaturePhoto, pk=photo_id)
+    feature = get_map_feature_or_404(feature_id, instance)
+    mf_photo = get_object_or_404(MapFeaturePhoto,
+                                 pk=photo_id,
+                                 map_feature=feature)
 
     image_data = mf_photo.image.read(settings.MAXIMUM_IMAGE_SIZE)
     mf_photo.set_image(image_data, degrees_to_rotate=degrees)
